pyproject.toml
[tool.poetry]
name = "pdf-threat-hunter"
version = "0.1.0"
description = "Multi-agent PDF static and threat-intel analysis toolkit"
authors = ["Gorelik"]
readme = "README.md"
packages = [{ include = "pdf_threat_hunter", from = "src" }]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"
langchain-core = "0.3.63"
langchain-openai = "0.3.18"
langgraph = "0.4.7"
langgraph-checkpoint = "2.0.26"
langgraph-prebuilt = "0.2.2"
langgraph-cli = { version = "0.0.19", extras = ["inmem"] }
openai = "1.82.1"
langsmith = { version = "0.3.43", optional = true }

pydantic = "^2.11"
pydantic-settings = "^2.2"
python-dotenv = "^1.1"

httpx = "^0.28"
tenacity = "^9.1"
tiktoken = "^0.9"
rich = "^13.7"
PyYAML = "^6.0.2"

Pillow = "^10.3"

[tool.poetry.extras]
tracing = ["langsmith"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-mock = "^3.12"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"